VesaBall v0.91old2
Snapshot dated 2005-01-14 23:14.
8 modified.

Imported from the original versioned zip archive.

+++ GameEditor.h
@@ -16,5 +16,5 @@ public:
 
 protected:
 	CCursor*	pCursor;
-	BYTE		curType;
+	LONG		curType;
 };
